backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
pyodbc==5.2.0
mssql-django==1.5

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
python-multipart==0.0.9

# Redis Cache
redis[hiredis]==5.0.1

# Validation and Settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Development Tools
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
